# documentation/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: Task Management API
  version: 1.0.0
  description: API documentation for the Task Management Platform backend.
servers:
  - url: http://localhost:8000/api
    description: Local development server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 2
        name:
          type: string
          example: Miss Maryam Gottlieb
        email:
          type: string
          format: email
          example: eunice65@example.com
        email_verified_at:
          type: string
          format: date-time
          nullable: true
          example: "2026-05-04T14:05:54.000000Z"
        role:
          type: string
          enum: [admin, manager, user]
          example: admin
        created_at:
          type: string
          format: date-time
          example: "2026-05-04T14:05:54.000000Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-05-04T14:05:54.000000Z"
    UserCompact:
      type: object
      description: Compact user object used inside task relations
      properties:
        id:
          type: integer
          example: 3
        name:
          type: string
          example: Mozelle Batz
        email:
          type: string
          format: email
          example: ratke.marlene@example.org
    Task:
      type: object
      properties:
        id:
          type: integer
          example: 12
        title:
          type: string
          example: Provident omnis in animi.
        description:
          type: string
          nullable: true
          example: Sit quia et est minima accusantium.
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
          example: in_progress
        priority:
          type: string
          enum: [low, medium, high]
          example: medium
        assigned_user_id:
          type: integer
          nullable: true
          example: 4
        created_by:
          type: integer
          example: 2
        due_date:
          type: string
          format: date-time
          nullable: true
          example: "2026-05-21T00:00:00.000000Z"
        created_at:
          type: string
          format: date-time
          example: "2026-05-04T14:05:55.000000Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-05-04T14:05:55.000000Z"
        assignee:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/UserCompact'
        creator:
          allOf:
            - $ref: '#/components/schemas/UserCompact'
    TaskDetail:
      allOf:
        - $ref: '#/components/schemas/Task'
        - type: object
          properties:
            attachments:
              type: array
              items:
                $ref: '#/components/schemas/Attachment'
            comments:
              type: array
              items:
                $ref: '#/components/schemas/Comment'
    Attachment:
      type: object
      properties:
        id:
          type: integer
          example: 1
        task_id:
          type: integer
          example: 16
        file_name:
          type: string
          example: Candidate-Instructions.md
        file_path:
          type: string
          example: attachments/m5nlZPoROMLXcTyrSvXln30ZQskSbU7izTj5ZjMN.txt
        file_size:
          type: integer
          example: 7206
        mime_type:
          type: string
          example: text/plain
        thumbnail_path:
          type: string
          nullable: true
          example: null
        uploaded_at:
          type: string
          format: date-time
          example: "2026-05-04T14:54:17.000000Z"
        created_at:
          type: string
          format: date-time
          example: "2026-05-04T14:54:17.000000Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-05-04T14:54:17.000000Z"
    Comment:
      type: object
      properties:
        id:
          type: integer
        task_id:
          type: integer
        user_id:
          type: integer
        comment:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/User'
    PaginationLinks:
      type: object
      properties:
        url:
          type: string
          nullable: true
        label:
          type: string
        page:
          type: integer
          nullable: true
        active:
          type: boolean
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
paths:
...
  /tasks:
    get:
      summary: List Tasks
      tags:
        - Tasks
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, in_progress, completed, cancelled]
        - in: query
          name: priority
          schema:
            type: string
            enum: [low, medium, high]
        - in: query
          name: sort_by
          schema:
            type: string
        - in: query
          name: sort_dir
          schema:
            type: string
            enum: [asc, desc]
        - in: query
          name: per_page
          schema:
            type: integer
            example: 10
        - in: query
          name: search
          schema:
            type: string
      responses:
        '200':
          description: Paginated list of tasks (Laravel paginator format)
          content:
            application/json:
              schema:
                type: object
                properties:
                  current_page:
                    type: integer
                    example: 1
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
                  first_page_url:
                    type: string
                    example: "http://localhost:8001/api/tasks?page=1"
                  from:
                    type: integer
                    example: 1
                  last_page:
                    type: integer
                    example: 2
                  last_page_url:
                    type: string
                    example: "http://localhost:8001/api/tasks?page=2"
                  links:
                    type: array
                    items:
                      $ref: '#/components/schemas/PaginationLinks'
                  next_page_url:
                    type: string
                    nullable: true
                    example: "http://localhost:8001/api/tasks?page=2"
                  path:
                    type: string
                    example: "http://localhost:8001/api/tasks"
                  per_page:
                    type: integer
                    example: 10
                  prev_page_url:
                    type: string
                    nullable: true
                    example: null
                  to:
                    type: integer
                    example: 10
                  total:
                    type: integer
                    example: 15
        '401':
          description: Unauthenticated

    post:
      summary: Create Task
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                  example: Example Task 1
                description:
                  type: string
                  example: This is example
                status:
                  type: string
                  enum: [pending, in_progress, completed, cancelled]
                  example: pending
                priority:
                  type: string
                  enum: [low, medium, high]
                  example: low
                assigned_user_id:
                  type: integer
                  example: 3
                due_date:
                  type: string
                  format: date
                  example: "2026-05-05"
      responses:
        '201':
          description: Task berhasil dibuat
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Task'
              example:
                message: "Task created successfully."
                data:
                  title: "Example Task 1"
                  description: "This is example"
                  status: "pending"
                  priority: "low"
                  assigned_user_id: "3"
                  due_date: "2026-05-05T00:00:00.000000Z"
                  created_by: 2
                  updated_at: "2026-05-04T14:23:32.000000Z"
                  created_at: "2026-05-04T14:23:32.000000Z"
                  id: 17
                  assignee:
                    id: 3
                    name: "Mozelle Batz"
                    email: "ratke.marlene@example.org"
                  creator:
                    id: 2
                    name: "Miss Maryam Gottlieb"
                    email: "eunice65@example.com"
        '422':
          description: Validation errors
        '401':
          description: Unauthenticated
  /tasks/{id}:
    get:
      summary: Get Task Details
      tags:
        - Tasks
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          example: 12
      responses:
        '200':
          description: Detail task beserta attachments dan comments
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/TaskDetail'
              example:
                data:
                  id: 12
                  title: "Provident omnis in animi."
                  description: "Sit quia et est minima accusantium. Enim explicabo quis ut sit. Ratione non repudiandae labore laborum."
                  status: "in_progress"
                  priority: "medium"
                  assigned_user_id: 4
                  created_by: 2
                  due_date: "2026-05-21T00:00:00.000000Z"
                  created_at: "2026-05-04T14:05:55.000000Z"
                  updated_at: "2026-05-04T14:05:55.000000Z"
                  assignee:
                    id: 4
                    name: "Prof. Loren Kilback"
                    email: "enos04@example.net"
                  creator:
                    id: 2
                    name: "Miss Maryam Gottlieb"
                    email: "eunice65@example.com"
                  attachments: []
                  comments: []
        '404':
          description: Task tidak ditemukan
        '401':
          description: Unauthenticated

    put:
      summary: Update Task
      tags:
        - Tasks
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          example: 12
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: Example Task 2
                description:
                  type: string
                  example: This is example
                status:
                  type: string
                  enum: [pending, in_progress, completed, cancelled]
                priority:
                  type: string
                  enum: [low, medium, high]
                assigned_user_id:
                  type: integer
                  example: 3
                due_date:
                  type: string
                  format: date
                  example: "2026-05-07"
      responses:
        '200':
          description: Task berhasil diupdate
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Task'
              example:
                message: "Task updated successfully."
                data:
                  id: 12
                  title: "Example Task 1"
                  description: "This is example"
                  status: "pending"
                  priority: "low"
                  assigned_user_id: "3"
                  created_by: 2
                  due_date: "2026-05-07T00:00:00.000000Z"
                  created_at: "2026-05-04T14:05:55.000000Z"
                  updated_at: "2026-05-04T14:25:38.000000Z"
                  assignee:
                    id: 3
                    name: "Mozelle Batz"
                    email: "ratke.marlene@example.org"
                  creator:
                    id: 2
                    name: "Miss Maryam Gottlieb"
                    email: "eunice65@example.com"
        '404':
          description: Task tidak ditemukan
        '422':
          description: Validation errors
        '401':
          description: Unauthenticated

    delete:
      summary: Delete Task
      tags:
        - Tasks
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          example: 16
      responses:
        '200':
          description: Task berhasil dihapus
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: "Task deleted successfully."
        '404':
          description: Task tidak ditemukan
        '401':
          description: Unauthenticated
...